update, insert 구문 개행문자 modifyPro file 값주기 No. index 값 job 게시판/댓글 수정삭제 notice 댓글 수정삭제

+++ The_First_modify/src/academy_community/action/AcademyDeleteProAction.java
@@ -8,9 +8,12 @@ import javax.servlet.http.HttpServletResponse;
 import academy_community.svc.AcademyCommentDeleteProService;
 import academy_community.svc.AcademyCommentListService;
 import academy_community.svc.AcademyDeleteProService;
+import academy_community.svc.AcademyDetailService;
 import academy_community.svc.AcademyModifyProService;
+import academy_community.vo.AcademyBean;
 import academy_community.vo.ActionForward;
 import any_community.svc.CommentDeleteProService;
+import member.svc.MemberUpdateProService;
 
 public class AcademyDeleteProAction implements Action {
 
@@ -27,6 +30,8 @@ public class AcademyDeleteProAction implements Action {
 			nowPage = "1";
 		}
 		
+		boolean isCommentDeleteSuccess =false;
+		
 		// 1)post_num에 대한 댓글 카운팅 
 		// 2) 댓글이 존재할경우 CommentDeleteProService에서 post_num으로 댓글 전부삭제
 		// 3) 댓글 전부 삭제가 완료될겅유 communityDeleteProService.deleteArticle 실행
@@ -35,23 +40,32 @@ public class AcademyDeleteProAction implements Action {
 			
 		if (commentListCount > 0) {
 			AcademyCommentDeleteProService academyCommentDeleteProService = new AcademyCommentDeleteProService();
-			boolean isCommentDeleteSuccess = academyCommentDeleteProService.deleteAllComment(num);
+			isCommentDeleteSuccess = academyCommentDeleteProService.deleteAllComment(num);
 		}
+		AcademyDetailService academyDetailService = new AcademyDetailService();
+		AcademyBean academyBean = academyDetailService.getArticle(num);
+		String nickname = academyBean.getNickname();
 		 
-		AcademyDeleteProService academyDeleteProService = new AcademyDeleteProService();
-		boolean isArticleDeleteSuccess = academyDeleteProService.removeArticle(num);
+		if(isCommentDeleteSuccess) {
+			AcademyDeleteProService academyDeleteProService = new AcademyDeleteProService();
+			boolean isArticleDeleteSuccess = academyDeleteProService.removeArticle(num);
 		
-		if (!isArticleDeleteSuccess) {
-			response.setContentType("text/html; charset=UTF-8");
-			PrintWriter out = response.getWriter();
-			out.println("<script>");
-			out.println("alert('글 삭제 실패!')");
-			out.println("history.back()");
-			out.println("</script>");
-		} else {
-			forward = new ActionForward();
-			forward.setPath("AcademyList.ac?page=" + nowPage);
-			forward.setRedirect(true);
+				if (!isArticleDeleteSuccess) {
+					response.setContentType("text/html; charset=UTF-8");
+					PrintWriter out = response.getWriter();
+					out.println("<script>");
+					out.println("alert('글 삭제 실패!')");
+					out.println("history.back()");
+					out.println("</script>");
+				} else {
+					MemberUpdateProService memberUpdateProService = new MemberUpdateProService();
+					boolean isSuccess = memberUpdateProService.minusArticletLP(nickname);
+					if(isSuccess) {
+						forward = new ActionForward();
+						forward.setPath("AcademyList.ac?page=" + nowPage);
+						forward.setRedirect(true);
+					}
+				}
 		}
 
 		return forward;
